rename: 커뮤니티 파일명 대소문자 임시 수정
