+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from "./errors";
 export * from "./mover.type";
 export * from "./notification.type";
 export * from "./request.type";
+export * from "./community.type.tmp";
